Wrap exceptions in subprocesses
Up until now, exceptions in subprocesses were still pickled, which will
fail if the exception is in a module that the host cannot import
directly.

This is the case when the path to the module gets dynamically added to the
python path in the code of the wrapped function. This is, for example,
the case for nest-simulations done in RunInSubprocess-wrapped functions.

Now all exceptions on the client side are transformed to strings and
then wrapped in a RemoteException as was originally intended but not
implemented.

Change-Id: I4ce80c52363d0a274bf19d40cea2454eb679c130